+++ forui/lib/src/widgets/resizable/resizable_controller.dart
@@ -1,10 +1,14 @@
-import 'dart:collection';
+import 'package:collection/collection.dart';
+import 'package:flutter/foundation.dart';
 
 import 'package:sugar/collection_aggregate.dart';
 
 import 'package:forui/forui.dart';
 import 'package:forui/src/foundation/doubles.dart';
 import 'package:forui/src/widgets/resizable/resizable_region_data.dart';
+
+part 'resizable_control.dart';
+part 'resizable_controller.control.dart';
 
 /// A controller that manages the resizing of regions in a [FResizable].
 abstract interface class FResizableController extends FChangeNotifier {