+++ forui/lib/widgets/resizable.dart
@@ -7,6 +7,6 @@ library forui.widgets.resizable;
 
 export '../src/widgets/resizable/resizable.dart' hide InheritedData;
 export '../src/widgets/resizable/divider.dart' hide Divider, HorizontalDivider, VerticalDivider;
-export '../src/widgets/resizable/resizable_controller.dart';
+export '../src/widgets/resizable/resizable_controller.dart' hide InternalFResizableControl;
 export '../src/widgets/resizable/resizable_region.dart';
 export '../src/widgets/resizable/resizable_region_data.dart' hide UpdatableResizableRegionData;